ai-backend/pyproject.toml
[tool.poetry]
name = "harmony-ai"
version = "0.1.1"
description = "A conversation AI chatbot for mental health, designed to provide support and act as a virtual therapist."
authors = ["Aryan Prince <hello@aryanprince.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11.6"
python-dotenv = "^1.0.1"
requests = "^2.32.2"
langchain = "^0.2.1"
langchain-groq = "^0.1.4"
deepgram-sdk = "^3.2.7"
pyaudio = "^0.2.14"
prettytable = "^3.10.0"
elevenlabs = "^1.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
